requirements.txt
Django==1.11
djangorestframework==3.6.2
Scrapy==1.3.3
beautifulsoup4==4.5.3
drfdocs=0.0.11
